% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCommonCluster.R
\name{calcCommonCluster}
\alias{calcCommonCluster}
\title{calcCommonCluster}
\usage{
calcCommonCluster(
  fitList,
  mergeThresh = 0,
  Fst = NULL,
  mxThresh = 0.01,
  Qallele = "99",
  returnPrepared = FALSE
)
}
\arguments{
\item{fitList}{List with fitted MLE-objects returned from a EuroForMix family program: List with sample elements which are again list elements: fitList[[sample]]=list(MLE,DC)}

\item{mergeThresh}{Threshold used for terminating the clustering: Note that this is relative LR}

\item{Fst}{The Fst used for the prior can be modified (over-writes existing)}

\item{mxThresh}{Threshold of whether multiple unknown contributors in an evididence should be merged}

\item{Qallele}{Name of compound drop-out alleles}

\item{returnPrepared}{Whether returning the pre-processed objects}
}
\description{
Provides clustering of unknown contributors in samples
}
\details{
Currently supported EuroForMix Family packages: (euroformix, EFMmps)

Implements the FAC-clustering approach (greedy-search) from 
Bhembe et al (2026): "Salience, legitimacy, and credibility of two end-to-end forensic single cell probabilistic systems"

Implementation structure
1. Validate input
2. Extract sample/marker/component information
3. Harmonize genotype spaces and construct global priors
4. Convert deconvolution posteriors to likelihood weights
5. Initialize singleton clusters
6. Calculate initial pairwise merge likelihoods
7. Greedily merge, reusing unaffected calculations
8. Return clusters
}
\author{
Oyvind Bleka
}
