% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcGenoProb.R
\name{calcGenoProb}
\alias{calcGenoProb}
\title{calcGenoProb}
\usage{
calcGenoProb(freq, nU = 1, fst = 0, nTyped = NULL)
}
\arguments{
\item{freq}{A vector of allele frequencies for a given population.}

\item{nU}{Number of unknowns}

\item{fst}{Assumed theta/fst-correction}

\item{nTyped}{contains a vector indicating number of typed allele}
}
\value{
Glist A list with genotypes and genotype probabilities
}
\description{
Returns a list of joint genotypes with corresponding joing probabilities for unknowns contributors for a given marker.
}
\details{
A simplified version of calcGjoint function from euroformix package
The function returns the list of all possible genotypes, with corresponding probabilities. The allele-names in popFreq needs to be numbers.
}
\author{
Oyvind Bleka
}
